+++ config/routing.yml
@@ -1,13 +1,17 @@
 ideas_index_controller:
     pattern: /ideas
-    defaults: { _controller: phpbb.ideas.controller:index }
+    defaults: { _controller: phpbb.ideas.index_controller:index }
 
 ideas_idea_controller:
     pattern: /idea/{idea_id}
-    defaults: { _controller: phpbb.ideas.controller:idea }
+    defaults: { _controller: phpbb.ideas.idea_controller:idea }
     requirements:
         idea_id: \d+
 
 ideas_list_controller:
     pattern: /ideas/all/{sort}
-    defaults: { _controller: phpbb.ideas.controller:all, sort: new }
+    defaults: { _controller: phpbb.ideas.list_controller:all, sort: new }
+
+ideas_post_controller:
+    pattern: /ideas/post
+    defaults: { _controller: phpbb.ideas.post_controller:post }
